+++ src/lib/checkbox/types.ts
@@ -3,5 +3,5 @@ import type { HTMLInputAttributes } from 'svelte/elements'
 export interface CheckboxProps extends Omit<HTMLInputAttributes, 'type'> {
 	element?: HTMLElement
 	group?: (string | number)[] | null
-	error?: boolean
+	issues?: { message: string }[]
 }